Login 2/Login/backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
fastapi
uvicorn
python-dotenv
pyodbc
sqlalchemy
passlib[bcrypt]
python-jose
